+++ xclim/indices/run_length.py
@@ -755,19 +755,58 @@ def extract_events(
     return runs
 
 
-def season(
+def season_start(
     da: xr.DataArray,
     window: int,
     date: DayOfYearStr | None = None,
     dim: str = "time",
-    stat: str | None = None,
     coord: str | bool | None = False,
-) -> xr.Dataset | xr.DataArray:
-    """Calculate the bounds of a season along a dimension.
+) -> xr.DataArray:
+    """Start of a season.
 
-    A "season" is a run of True values that may include breaks under a given length (`window`).
-    The start is computed as the first run of `window` True values, then end as the first subsequent run
-    of  `window` False values. If a date is passed, it must be included in the season.
+    See :py:func:`season`.
+
+    Parameters
+    ----------
+    da : xr.DataArray
+        Input N-dimensional DataArray (boolean).
+    window : int
+        Minimum duration of consecutive values to start and end the season.
+    date : DayOfYearStr, optional
+        The date (in MM-DD format) that a season must include to be considered valid.
+    dim : str
+        Dimension along which to calculate the season (default: 'time').
+    coord : Optional[str]
+        If not False, the function returns values along `dim` instead of indexes.
+        If `dim` has a datetime dtype, `coord` can also be a str of the name of the
+        DateTimeAccessor object to use (ex: 'dayofyear').
+
+    Returns
+    -------
+    xr.DataArray
+        Start of the season, units depend on `coord`.
+
+    See Also
+    --------
+    season
+    season_end
+    season_length
+    """
+    return first_run_before_date(da, window=window, date=date, dim=dim, coord=coord)
+
+
+def season_end(
+    da: xr.DataArray,
+    window: int,
+    date: DayOfYearStr | None = None,
+    dim: str = "time",
+    coord: str | bool | None = False,
+    _beg: xr.DataArray | None = None,
+) -> xr.DataArray:
+    """End of a season.
+
+    See :py:func:`season`. There are two differences between this and :py:func:`first_run_after_date`:
+
 
     Parameters
     ----------
@@ -779,9 +818,6 @@ def season(
         The date (in MM-DD format) that a run must include to be considered valid.
     dim : str
         Dimension along which to calculate consecutive run (default: 'time').
-    stat : {'start', 'end', 'length'}, optional
-        Which facet to return. If absent (default), a Dataset with all three is returned.
-        If present, only the specified variable is returned.
     coord : Optional[str]
         If not False, the function returns values along `dim` instead of indexes.
         If `dim` has a datetime dtype, `coord` can also be a str of the name of the
@@ -789,9 +825,71 @@ def season(
 
     Returns
     -------
-    xr.Dataset or xr.DataArray
-        If `stat` is absent, a Dataset with three variables : start, end and length of the season.
-        If `stat` is given, only the specified variable is returned, as a DataArray.
+    xr.DataArray
+        End of the season, units depend on `coord`.
+        If there is a start is found but no end, the end is set to the last element.
+
+    See Also
+    --------
+    season
+    season_start
+    season_length
+    """
+    # Fast path for `season` and `season_length`
+    if _beg is not None:
+        beg = _beg
+    else:
+        beg = season_start(da, window=window, dim=dim, date=date, coord=False)
+    # Invert the condition and mask all values after beginning
+    # we fillna(0) as so to differentiate series with no runs and all-nan series
+    not_da = (~da).where(da[dim].copy(data=np.arange(da[dim].size)) >= beg.fillna(0))
+    end = first_run_after_date(not_da, window=window, dim=dim, date=date, coord=False)
+    if _beg is None:
+        # Where end is null and beg is not null (valid data, no end detected), put the last index
+        # Don't do this in the fast path, so that the length can use this information
+        end = xr.where(end.isnull() & beg.notnull(), da[dim].size - 1, end)
+        end = end.where(beg.notnull())
+    if coord:
+        crd = da[dim]
+        if isinstance(coord, str):
+            crd = getattr(crd.dt, coord)
+        end = lazy_indexing(crd, end)
+    return end
+
+
+def season(
+    da: xr.DataArray,
+    window: int,
+    date: DayOfYearStr | None = None,
+    dim: str = "time",
+    coord: str | bool | None = False,
+) -> xr.Dataset:
+    """Calculate the bounds of a season along a dimension.
+
+    A "season" is a run of True values that may include breaks under a given length (`window`).
+    The start is computed as the first run of `window` True values, and the end as the first subsequent run
+    of  `window` False values. The end element is the first element after the season.
+    If a date is given, it must be included in the season.
+
+    Parameters
+    ----------
+    da : xr.DataArray
+        Input N-dimensional DataArray (boolean).
+    window : int
+        Minimum duration of consecutive values to start and end the season.
+    date : DayOfYearStr, optional
+        The date (in MM-DD format) that a run must include to be considered valid.
+    dim : str
+        Dimension along which to calculate consecutive run (default: 'time').
+    coord : Optional[str]
+        If not False, the function returns values along `dim` instead of indexes.
+        If `dim` has a datetime dtype, `coord` can also be a str of the name of the
+        DateTimeAccessor object to use (ex: 'dayofyear').
+
+    Returns
+    -------
+    xr.Dataset
+        A Dataset with three variables : start, end and length of the season.
 
     Notes
     -----
@@ -803,55 +901,36 @@ def season(
     25 for all June, but July and august have very cold temperatures. Instead of returning 30 days (June), the function
     will return 61 days (July + June).
 
-    The season's length is always the difference between the end and the start, except if no
-    end was found before the end of the data, in which case the end is the last day and the length
-    goes up to the boundary (i.e. :math:`end - start = length - 1`).
+    The season's length is always the difference between the end and the start. Except if no
+    season end was found before the end of the data. In that case the end is set to last element and
+    the length is set to the data size minus the start index. Thus, for the specific case, :math:`length = end - start + 1`,
+    because the end falls on the last element of the season instead of the subsequent one.
+
+    See Also
+    --------
+    season_start
+    season_end
+    season_length
     """
-    beg = first_run(da, window=window, dim=dim)
-    # Invert the condition and mask all values after beginning
-    # we fillna(0) as so to differentiate series with no runs and all-nan series
-    not_da = (~da).where(da[dim].copy(data=np.arange(da[dim].size)) >= beg.fillna(0))
-
-    # Mask also values after "date"
-    mid_idx = index_of_date(da[dim], date, max_idxs=1, default=0)
-    if mid_idx.size == 0:
-        # The date is not within the group. Happens at boundaries.
-        base = da.isel({dim: 0})  # To have the proper shape
-        beg = xr.full_like(base, np.nan, float).drop_vars(dim)
-        end = xr.full_like(base, np.nan, float).drop_vars(dim)
-        length = xr.full_like(base, np.nan, float).drop_vars(dim)
-    else:
-        if date is not None:
-            # If the beginning was after the mid date, both bounds are NaT.
-            valid_start = beg < mid_idx.squeeze()
-        else:
-            valid_start = True
-
-        not_da = not_da.where(da[dim] >= da[dim][mid_idx][0])
-        end = first_run(
-            not_da,
-            window=window,
-            dim=dim,
-        )
-        # If there was a beginning but no end, season goes to the end of the array
-        no_end = beg.notnull() & end.isnull()
-
-        # Length
-        length = end - beg
-
-        # No end:  length is actually until the end of the array, so it is missing 1
-        length = xr.where(no_end, da[dim].size - beg, length)
-        # Where the beginning was before the mid-date, invalid.
-        length = length.where(valid_start)
-        # Where there were data points, but no season : put 0 length
-        length = xr.where(beg.isnull() & end.notnull(), 0, length)
-
-        # No end: end defaults to the last element (this differs from length, but heh)
-        end = xr.where(no_end, da[dim].size - 1, end)
-
-        # Where the beginning was before the mid-date
-        beg = beg.where(valid_start)
-        end = end.where(valid_start)
+    beg = season_start(da, window=window, dim=dim, date=date, coord=False)
+    # Use fast path in season_end : no recomputing of start, no masking of end where beg.isnull() and don't set end if none found
+    end = season_end(da, window=window, dim=dim, date=date, _beg=beg, coord=False)
+    # Three cases :
+    #           start     no start
+    # end       e - s        0
+    # no end   size - s      0
+    # da is boolean, so we have no way of knowing if the absence of season is due to missing values or to an actual absence.
+    length = xr.where(
+        beg.isnull(),
+        0,
+        # Where there is no end, from the start to the boundary
+        xr.where(end.isnull(), da[dim].size - beg, end - beg),
+    )
+    # Now masks ends
+    # Still give an end if we didn't find any : put the last element
+    # This breaks the length = end - beg, but yields a truer length
+    end = xr.where(end.isnull() & beg.notnull(), da[dim].size - 1, end)
+    end = end.where(beg.notnull())
 
     if coord:
         crd = da[dim]
@@ -866,7 +945,6 @@ def season(
         coordstr = "index"
 
     out = xr.Dataset({"start": beg, "end": end, "length": length})
-
     out.start.attrs.update(
         long_name="Start of the season.",
         description=f"First {coordstr} of a run of at least {window} steps respecting the condition.",
@@ -880,8 +958,6 @@ def season(
         long_name="Length of the season.",
         description="Number of steps of the original series in the season, between 'start' and 'end'.",
     )
-    if stat is not None:
-        return out[stat]
     return out
 
 
@@ -891,11 +967,9 @@ def season_length(
     date: DayOfYearStr | None = None,
     dim: str = "time",
 ) -> xr.DataArray:
-    """Return the length of the longest semi-consecutive run of True values (optionally including a given date).
+    """Length of a season.
 
-    A "season" is a run of True values that may include breaks under a given length (`window`).
-    The start is computed as the first run of `window` True values, then end as the first subsequent run
-    of  `window` False values. If a date is passed, it must be included in the season.
+    See :py:func:`season`.
 
     Parameters
     ----------
@@ -911,18 +985,13 @@ def season_length(
     Returns
     -------
     xr.DataArray, [int]
-        Length of the longest run of True values along a given dimension (inclusive of a given date)
-        without breaks longer than a given length.
+        Length of the season, in number of elements along dimension `time`.
 
-    Notes
-    -----
-    The run can include holes of False or NaN values, so long as they do not exceed the window size.
-
-    If a date is given, the season start and end are forced to be on each side of this date. This means that
-    even if the "real" season has been over for a long time, this is the date used in the length calculation.
-    Example : Length of the "warm season", where T > 25°C, with date = 1st August. Let's say the temperature is over
-    25 for all June, but July and august have very cold temperatures. Instead of returning 30 days (June), the function
-    will return 61 days (July + June).
+    See Also
+    --------
+    season
+    season_start
+    season_end
     """
     seas = season(da, window, date, dim, coord=False)
     return seas.length
@@ -1062,6 +1131,54 @@ def last_run_before_date(
 
     run = da.where(da[dim] <= da[dim][mid_idx][0])
     return last_run(run, window=window, dim=dim, coord=coord)
+
+
+def first_run_before_date(
+    da: xr.DataArray,
+    window: int,
+    date: DayOfYearStr | None = "07-01",
+    dim: str = "time",
+    coord: bool | str | None = "dayofyear",
+) -> xr.DataArray:
+    """Return the index of the first item of the first run before a given date.
+
+    Parameters
+    ----------
+    da : xr.DataArray
+        Input N-dimensional DataArray (boolean).
+    window : int
+        Minimum duration of consecutive run to accumulate values.
+    date : DayOfYearStr
+        The date before which to look for the run.
+    dim : str
+        Dimension along which to calculate consecutive run (default: 'time').
+    coord : Optional[Union[bool, str]]
+        If not False, the function returns values along `dim` instead of indexes.
+        If `dim` has a datetime dtype, `coord` can also be a str of the name of the
+        DateTimeAccessor object to use (ex: 'dayofyear').
+
+    Returns
+    -------
+    xr.DataArray
+        Index (or coordinate if `coord` is not False) of first item in the first valid run.
+        Returns np.nan if there are no valid runs.
+    """
+    if date is not None:
+        mid_idx = index_of_date(da[dim], date, max_idxs=1, default=0)
+        if (
+            mid_idx.size == 0
+        ):  # The date is not within the group. Happens at boundaries.
+            return xr.full_like(da.isel({dim: 0}), np.nan, float).drop_vars(dim)
+        # Mask anything after the mid_date + window - 1
+        # Thus, the latest run possible can begin on the day just before mid_idx
+        da = da.where(da[dim] < da[dim][mid_idx + window - 1][0])
+
+    return first_run(
+        da,
+        window=window,
+        dim=dim,
+        coord=coord,
+    )
 
 
 @njit